refactor: make one folder for all components
