+++ app/signin/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import Signin from "../components/common/Signin";
+import Signin from "../../components/common/Signin";
 import { authOptions } from "../api/auth/[...nextauth]/route";
 import { redirect } from "next/navigation";
